refactor(path): Split up mod
Cherry pick ef2b41e7759cada65d91cd61addadc8db2786af8 (#297)
